# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "1.0.0"
  title: Shoe Order API
# during dev, should point to your local machine
host: localhost:10010
...
basePath: /api/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
  - application/x-www-form-urlencoded
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /customers:
    x-swagger-router-controller: customer_controller
    get:
      tags: ["customers"]
      summary: Returns all created customers with the basic informations
      operationId: getAllCustomers
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/GetCustomersListResponse"
    post:
      tags: ["customers"]
      summary: Create a new customer
      operationId: createCustomer
      parameters:
        - name: username
          description: Can be anything you want
          in: formData
          required: true
          type: string
        - name: password
          description: At least 8 chars
          in: formData
          required: true
          type: string
        - name: firstname
          description: Firstame of the new customer
          in: formData
          required: true
          type: string
        - name: lastname
          description: Lastname of the new customer
          in: formData
          required: true
          type: string
        - name: email
          description: Email of the new customer
          in: formData
          required: true
          type: string
        - name: birthdate
          description: Birthdate of the new customer
          in: formData
          required: true
          type: string
          format: date
        - name: current_address
          description: Current address of the customer
          in: formData
          required: false
          type: integer
        - name: is_active
          description: TRUE if customer is active, else FALSE
          in: formData
          required: true
          type: boolean
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/PostCustomerResponse"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/PostCustomerBadRquestError"
  /customers/{customerId}:
    x-swagger-router-controller: customer_controller
    get:
      tags: ["customers"]
      summary: Get a customer by his ID
      description: Returns a customer by his ID with all basic informations
      operationId: getCustomerById
      parameters:
        - name: customerId
          type: string
          in: path
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/GetCustomerResponse"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/Error"
    put:
      tags: ["customers"]
      summary: Modify all properties of a customer
      operationId: fullEditCustomer
      parameters:
        - name: customerId
          description: Id of the customer to update
          type: string
          in: path
          required: true
        - name: username
          description: Username of the customer to update
          in: formData
          required: true
          type: string
        - name: password
          description: Password of the customer to update
          in: formData
          required: true
          type: string
        - name: firstname
          description: Firstame of the customer to update
          in: formData
          required: true
          type: string
        - name: lastname
          description: Lastname of the customer to update
          in: formData
          required: true
          type: string
        - name: email
          description: Email of the customer to update
          in: formData
          required: true
          type: string
        - name: birthdate
          description: Birthdate of the customer to update
          in: formData
          required: true
          type: string
          format: date
        - name: current_address
          description: Current address of the customer
          in: formData
          required: false
          type: integer
        - name: is_active
          description: TRUE if customer is active, else FALSE
          in: formData
          required: true
          type: boolean
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/FullEditCustomerResponse"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/FullEditCustomerByIdNotFound"
    patch:
      tags: ["customers"]
      summary: Modify some properties of a customer
      operationId: partialEditCustomer
      parameters:
        - name: customerId
          description: Id of the customer to update
          type: string
          in: path
          required: true
        - name: username
          description: Username of the customer to update
          in: formData
          required: false
          type: string
        - name: password
          description: Password of the customer to update
          in: formData
          required: false
          type: string
        - name: firstname
          description: Firstame of the customer to update
          in: formData
          required: false
          type: string
        - name: lastname
          description: Lastname of the customer to update
          in: formData
          required: false
          type: string
        - name: email
          description: Email of the customer to update
          in: formData
          required: false
          type: string
        - name: birthdate
          description: Birthdate of the customer to update
          in: formData
          required: false
          type: string
          format: date
        - name: current_address
          description: Current address of the customer
          in: formData
          required: false
          type: integer
        - name: is_active
          description: TRUE if customer is active, else FALSE
          in: formData
          required: true
          type: boolean
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/PartialEditCustomerResponse"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/PartialEditCustomerByIdNotFound"
    delete:
      tags: ["customers"]
      summary: Delete a specific customer with his Id
      description: Will delete all informations about a specific customer
      operationId: removeCustomer
      parameters:
        - name: customerId
          type: string
          in: path
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/DeleteCustomerByIdSuccessOrError"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/DeleteCustomerByIdSuccessOrError"
...
  /customizations:
    x-swagger-router-controller: customization_controller
    get:
      tags: ["customizations"]
      summary: Returns all created customization with the basic informations
      operationId: getAllCustomizations
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/GetCustomizationsListResponse"
    post:
      tags: ["customizations"]
      summary: Create a new customization
      operationId: createCustomization
      parameters:
        - name: customization
          description: The customization data
          in: body
          required: true
          schema:
            $ref: '#/definitions/CustomizationObject'
      responses:
        201:
          description: Created
          schema:
            $ref: "#/definitions/PostCustomizationResponse"
        400:
          description: Bad Request
          schema:
            $ref: "#/definitions/PostCustomizationBadRequestError"
  /customizations/{customizationId}:
    x-swagger-router-controller: customization_controller
    get:
      tags: ["customizations"]
      summary: Get a customization by his ID
      description: Returns a customization by his ID with all basic informations
      operationId: getCustomizationById
      parameters:
        - name: customizationId
          type: string
          in: path
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/GetCustomizationResponse"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/Error"
    patch:
      tags: ["customizations"]
      summary: Modify a customization by his ID
      description: Modify a customization by his ID and return the new object
      operationId: partialOrCompleteEditOfCustomization
      parameters:
        - name: customizationId
          description: The ID of the customization to modify
          type: string
          in: path
          required: true
        - name: customization
          description: The customization data
          in: body
          required: true
          schema:
            $ref: "#/definitions/CustomizationObject"
      responses:
        200:
          description: Success
          schema:
            $ref: "#/definitions/GeneralResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags: ["customizations"]
      summary: Delete a specific customization with his Id
      description: Will delete all informations about a specific customization
      operationId: deleteCustomizationById
      parameters:
        - name: customizationId
          type: string
          in: path
          required: true
      responses:
        200:
          description: OK
          schema:
            $ref: "#/definitions/DeleteCustomerByIdSuccessOrError"
        404:
          description: Not Found
          schema:
            $ref: "#/definitions/DeleteCustomerByIdSuccessOrError"

# complex objects have schema definitions
definitions:
  CustomizationObject:
    type: object
    required:
      - article_id
      - customer_id
      - customization
    properties:
      article_id:
        type: integer
        description: The ID of the article to customize
      customer_id:
        type: integer
        description: The ID of the customer who will create the customization
      customization:
        type: object
        properties:
          parts:
            type: array
            items:
              type: object
          size:
            type: string
...
  Error:
    type: object
    properties:
      code:
        type: number
      message:
        type: string
    required:
      - code
      - message
...
  GetCustomizationsListResponse:
    required:
      - code
      - customizations
    properties:
      code:
        type: integer
      customizations:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            article_id:
              type: integer
            customer_id:
              type: integer
            customization:
              type: string
            created_at:
              type: string
            updated_at:
              type: string
...
  GetCustomersListResponse:
    required:
      - code
      - customers
    properties:
      code:
        type: number
      customers:
        type: array
        items:
          type: object
          properties:
            id:
              type: integer
            firstname:
              type: string
            lastname:
              type: string
            email:
              type: string
            birthdate:
              type: string
              format: date
            current_address:
              type: integer
            is_active:
              type: boolean
            created_at:
              type: string
            updated_at:
              type: string
  GetCustomerResponse:
    required:
      - code
      - customer
    type: object
    properties:
      code:
        type: number
      customer:
        type: object
        properties:
          id:
            type: integer
          firstname:
            type: string
          lastname:
            type: string
          email:
            type: string
          birthdate:
            type: string
            format: date
          current_address:
            type: integer
          is_active:
            type: boolean
          created_at:
            type: string
          updated_at:
            type: string
  GetCustomizationResponse:
    required:
      - code
      - customization
    properties:
      code:
        type: integer
      customization:
        type: object
        properties:
          id:
            type: integer
          article_id:
            type: integer
          customer_id:
            type: integer
          customization:
            type: object
            properties:
              parts:
                type: array
                items:
                  type: object
              size:
                type: string
...
  PostCustomizationResponse:
    required:
      - code
      - customization
    properties:
      code:
        type: integer
      customization:
        type: object
        properties:
          article_id:
            type: integer
          customer_id:
            type: integer
          customization:
            type: object
            properties:
              parts:
                type: array
                items:
                  type: object
              size:
                type: string
  PostCustomizationBadRequestError:
    required:
      - code
      - type
    type: object
    properties:
      code:
        type: number
      type:
        type: string
      messages:
        type: array
        items:
          type: string
      fields:
        type: array
        items:
          type: string
...
  PostCustomerResponse:
    required:
      - code
      - customer
    properties:
      code:
        type: number
      customer:
        type: object
        properties:
          id:
            type: integer
          firstname:
            type: string
          lastname:
            type: string
          email:
            type: string
          birthdate:
            type: string
            format: date
          current_address:
            type: integer
          is_active:
            type: boolean
          created_at:
            type: string
          updated_at:
            type: string
  PostCustomerBadRquestError:
    required:
      - code
      - type
    type: object
    properties:
      code:
        type: number
      type:
        type: string
      messages:
        type: array
        items:
          type: string
      fields:
        type: array
        items:
          type: string
  FullEditCustomerResponse:
    required:
      - code
      - message
      - customer
    type: object
    properties:
      code:
        type: number
      message:
        type: string
      customer:
        type: object
        properties:
          id:
            type: integer
          firstname:
            type: string
          lastname:
            type: string
          email:
            type: string
          birthdate:
            type: string
            format: date
          current_address:
            type: integer
          is_active:
            type: boolean
          created_at:
            type: string
          updated_at:
            type: string
...
  PartialEditCustomerResponse:
    required:
      - code
      - message
      - customer
    type: object
    properties:
      code:
        type: number
      message:
        type: string
      customer:
        type: object
        properties:
          id:
            type: integer
          firstname:
            type: string
          lastname:
            type: string
          email:
            type: string
          birthdate:
            type: string
            format: date
          current_address:
            type: integer
          is_active:
            type: boolean
          created_at:
            type: string
          updated_at:
            type: string
  FullEditCustomerByIdNotFound:
    required:
      - code
      - message
    type: object
    properties:
      code:
        type: number
      message:
        type: string
  PartialEditCustomerByIdNotFound:
    required:
      - code
      - message
    type: object
    properties:
      code:
        type: number
      message:
        type: string
  DeleteCustomerByIdSuccessOrError:
    required:
      - code
      - message
    type: object
    properties:
      code:
        type: number
      message:
        type: string
...